requirements.txt
greenlet==2.0.1
psycopg2-binary==2.9.5
PyYAML==6.0
SQLAlchemy==1.4.46
